(REF) Extract ComposerIoTrait
