+++ src/Util/ShellRunner.php
@@ -4,33 +4,11 @@ namespace Civi\CompilePlugin\Util;
 
 use Composer\EventDispatcher\Event;
 use Composer\EventDispatcher\EventDispatcher;
-use Composer\IO\IOInterface;
-use Composer\Composer;
 
 class ShellRunner
 {
-    /**
-     * @var Composer
-     */
-    protected $composer;
 
-    /**
-     * @var IOInterface
-     */
-    protected $io;
-
-    /**
-     * ShellRunner constructor.
-     * @param \Composer\Composer $composer
-     * @param \Composer\IO\IOInterface $io
-     */
-    public function __construct(
-        \Composer\Composer $composer,
-        \Composer\IO\IOInterface $io
-    ) {
-        $this->composer = $composer;
-        $this->io = $io;
-    }
+    use ComposerIoTrait;
 
     /**
      * Run a shell command in the same style as Composer's EventDispatcher.
